# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Book Catalog API Documentation",
        "description": "Laravel Book Catalog API Documentation",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/rubrics": {
            "get": {
                "tags": [
                    "Rubrics"
                ],
                "summary": "Get list of rubrics",
                "description": "Returns list of rubrics",
                "operationId": "getRubricsList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {}
                ]
            },
            "post": {
                "tags": [
                    "Rubrics"
                ],
                "summary": "Store new rubric",
                "description": "Returns rubric data",
                "operationId": "storeRubric",
                "requestBody": {
                    "description": "Data packet for Test",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreRubricRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key_security": []
                    }
                ]
            }
        },
        "/rubrics/{id}": {
            "get": {
                "tags": [
                    "Rubrics"
                ],
                "summary": "Get rubric information",
                "description": "Returns rubric data",
                "operationId": "getRubricById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Rubric id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {}
                ]
            },
            "delete": {
                "tags": [
                    "Rubrics"
                ],
                "summary": "Delete rubric",
                "description": "Delete rubric",
                "operationId": "deleteRubric",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Rubric id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key_security": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Rubrics"
                ],
                "summary": "Edit rubric",
                "description": "Returns edited rubric data",
                "operationId": "editRubric",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Rubric id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data packet for Test",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreRubricRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key_security": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "StoreRubricRequest": {
                "title": "StoreRubricRequest",
                "description": "Store rubric request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Rubric`s name",
                        "type": "string",
                        "items": {
                            "type": "string",
                            "items": {}
                        },
                        "example": "Rubric 5"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "api_key_security": {
                "type": "apiKey",
                "description": " Enter api_token in format < Bearer your_token>",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "api_key_security": []
        }
    ]
}